requirements.txt
# Niagara BAS Downloader v2.0
# pip install -r requirements.txt

requests>=2.28.0
urllib3>=1.26.0
python-dotenv>=0.19.0
python-dateutil>=2.8.0
selenium>=4.0.0
customtkinter>=5.0.0
